Add Recurse Center OAuth log in
Integrate with the Recurse Center OAuth service to allow for protected
resources.

As a first step, allow full access to the index page, but display
different content depending on whether or not the user is logged in;
also introduce a `/secure` endpoint that requires OAuth login and prints
out the current user's name.

Although my goal is to eventually allow users to set up this project in
an environment outside of RC, for the time being it's more effective to
have some hardcoded data. For now, that's confined to the application
configuration file; I'll try to keep such RC-specific code contained in
that file.

Issue #2 Set up RC Authentication