Log RC profile details on successful login
Start setting up the process to persist a Recurser's profile to the
database after successfully logging in, by getting the information we
care about (at least, a first estimation of the information we care
about) out into its own POJO. Use lombok[1] to automatically generate
the various boilerplate for the POJO, as well as to automatically
construct a logger. Use Jackson[2] to assign the data from the
automatically-fetched API response to the POJO.

This pretty much marks my abandonment of attempting to constrain
RC-specific things to configuration; I realized that several of my
planned features depend on the RC API, so I might as well lean into it.

[1] https://projectlombok.org/
[2] https://github.com/FasterXML/jackson

Issue #3 Create author page