Show descriptive error if missing env variables
When Spring tries to configure itself when some of the environment
variables we reference in application.yml are missing, the error
messages can be long and obscure. Create a list of required environment
variables, check for them before entering Spring, and show a descriptive
error message if any are missing.