Simplify getting the currently logged in user
Many of the controller methods will need to access the currently logged
in user. Create a `CurrentUser` annotation which can be used on a `User`
controller method parameter.

This works by wrapping the OAuth2 service to do some post-processing.
See the Spring Security reference docs on how to specify the alternative
service[1], how the `AuthenticationPrincipal` annotation works
(including the suggestion to make a `CurrentUser` annotation!)[2], and
the documentation on the Spring Expression Language (SpEL)[3], as used
within the `AuthenticationPrincipal` expression.

[1] https://docs.spring.io/spring-security/site/docs/current/reference/html5/#oauth2login-advanced-userinfo-endpoint
[2] https://docs.spring.io/spring-security/site/docs/current/reference/html5/#mvc-authentication-principal
[3] https://docs.spring.io/spring-framework/docs/current/spring-framework-reference/core.html#expressions-language-ref

Issue #2 Set up RC Authentication
Issue #3 Create author page