Add user profiles and visibility policy
Create the suite of user profile pages: viewing your profile, editing
your profile, viewing someone else's profile as a logged-in user, and
viewing someone else's profile as a logged-out user.

As part of that, create a framework for setting and checking a
visibility policy for any given thing. For now, this is just used for
profile visibility.

Issue #3 Create author page