Redirect to previously viewed page after login
Rather than redirect to the homepage, make the login link add a query
parameter to the current page, and require authentication for URLs
ending with that query parameter.

This has the drawback of leaving a `?login` parameter in the URL, but
for now the simplicity of configuration and the difficulty of the
alternatives is worth that drawback.