Add pages to show & edit (but not create) projects
Create new models, controllers, and tables to support viewing and
editing projects. Creation will be added separately.

I was going to defer having multiple authors on a project, and changing
the authors is not yet implemented, but due to the way Spring Data JDBC
works, set up the multiple-author infrastructure now. Specifically,
Spring Data JDBC does not support many-to-many relationships[1], and it
wants to delete and recreate referenced entities on edit, so rather than
a project having an author, follow the suggestion from the blog post to
have a class that just has the author ID, and augment that with a custom
query method on the repository to avoid making N+1 queries.

As part of this, generify the exception thrown when something (formerly
a user) is not found. Likewise, allow for multiple authors in a
visibility policy evaluation.

[1] https://spring.io/blog/2018/09/24/spring-data-jdbc-references-and-aggregates

Issue #4 Allow users to create projects