Allow users to create projects
Add "new project" GET and POST routes, and save the new project to the
database.

It turns out Spring Data JDBC seems to *really* dislike relational data.
I wasn't able to get the suggestion[1] I followed earlier to have a
placeholder class for the many-to-many relation to work with inserting
data; it wasn't populating the reference to the project.

The upside to this ORM is that it makes it relatively easy to just write
SQL. Manage the relationship by hand, rather than automatically, and
annotate the controller method as transactional to avoid partial
updates.

[1] https://spring.io/blog/2018/09/24/spring-data-jdbc-references-and-aggregates

Issue #4 Allow users to create projects