Add CSS to base layout page
Style the header to be a bar, with left and right-aligned content, a la
GitHub or Stack Overflow or any number of other modern websites. Set the
username to be last on the list, even though in a no-style context I
think it belongs before the create new project link, because I think it
looks better and that seems to be the convention.

Style the footer to be at the bottom of the page - even if there is less
than a screen's worth of content - and centered.

Issue #10 Improve appearance of pages