Style user pages
Add styling to the various user pages to make them look a little nicer.

Also add a placeholder image for a user's profile picture. So far, we
use this only in the raw HTML, and it is always replaced when rendered.

The placeholder image is named Portrait placeholder[1], by
Andomedium[2], licensed CC0.

[1] https://en.wikipedia.org/wiki/File:Portrait_placeholder.png
[2] https://en.wikipedia.org/wiki/User:Andomedium

Issue #10 Improve appearance of pages