Fix user creation by properly (un)setting user ID
In commit 0d4efa03549f9b112f1277502424876bf2238c27, I changed the `User`
model from using an `int` to an `Integer` for the `userId`. Spring Data
JDBC looks at the `Id`-annotated field to determine if it should update
or create, and while `int(0)` indicates a new entity, `Integer(0)`
indicates an existing entity.

Set the `userId` to `null` to create a new user.