Add a default public profile picture
Set the public image for new profiles to the placeholder user image,
which was originally only used for layout. That way, if and when they
set their profile to be public, the default behavior is to have some
kind of image rather than none at all.

Issue #43 Use placeholder image for new users