Require project and user names
Validate the input sent to the server to make sure that the user's
public and internal names are present and not empty, and to make sure
that the project's name is present and not empty.

Mark the fields as `required`[1] in the HTML so that the browser will
help us out.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#attr-required

Issue #42 Validate user input

Co-authored-by: Travis Horton <travis@travish.com>