Get the correct number of multi-author projects
Fix a lurking not-yet-bug introduced in commit
5bdb236648ddac7c18ad27e5cd9935be07c4cba1 where projects with multiple
authors would cause fewer projects to be fetched. In the process,
extract a shared SQL string for the common bits of the query.

This is relevant now because adding the ability to browse projects by
tag depends on browsing projects with pagination, and that'll in turn
require updating these queries.

Issue #5 Add tags to projects
Issue #32 Browse projects