Collapse most project description inputs
To make it a little less daunting to create a new project, hide the
internal and private project descriptions by default. Likewise, when
editing a project, hide any description fields that are empty. The
fields can be expanded or collapsed by clicking on the titles.

Issue #45 De-emphasize private project description